+++ packages/extension-polkagate/src/fullscreen/homeFullScreen/partials/AccountInformationForHome.tsx
@@ -5,24 +5,26 @@
 /* eslint-disable react/jsx-max-props-per-line */
 
 import type { BalancesInfo } from '@polkadot/extension-polkagate/util/types';
-import type { BN } from '@polkadot/util';
 import type { HexString } from '@polkadot/util/types';
 import type { FetchedBalance } from '../../../hooks/useAssetsBalances';
+import type { ItemInformation } from '../../nft/utils/types';
 
 import { ArrowForwardIos as ArrowForwardIosIcon, MoreVert as MoreVertIcon } from '@mui/icons-material';
 import { Box, Button, Divider, Grid, Typography, useTheme } from '@mui/material';
-import React, { useCallback, useContext, useMemo, useState } from 'react';
+import React, { useCallback, useEffect, useMemo, useState } from 'react';
 
 import { getValue } from '@polkadot/extension-polkagate/src/popup/account/util';
+import { type BN, noop } from '@polkadot/util';
 
 import { stars6Black, stars6White } from '../../../assets/icons';
-import { ActionContext } from '../../../components';
+import NftManager from '../../../class/nftManager';
 import FormatPrice from '../../../components/FormatPrice';
 import { useAccount, useCurrency, usePrices, useTranslation } from '../../../hooks';
 import { tieAccount } from '../../../messaging';
 import { amountToHuman } from '../../../util/utils';
 import AOC from '../../accountDetails/components/AOC';
 import { openOrFocusTab } from '../../accountDetails/components/CommonTasks';
+import NftGrouped from '../../accountDetails/components/NftGrouped';
 import DeriveAccountModal from '../../partials/DeriveAccountModal';
 import ExportAccountModal from '../../partials/ExportAccountModal';
 import ForgetAccountModal from '../../partials/ForgetAccountModal';
@@ -40,7 +42,12 @@ interface AddressDetailsProps {
   isChild?: boolean;
 }
 
-interface AccountButtonType { text: string, onClick: () => void, icon: React.ReactNode }
+interface AccountButtonType {
+  text: string;
+  onClick: () => void;
+  icon: React.ReactNode;
+  collapse?: boolean;
+}
 
 export enum POPUPS_NUMBER {
   DERIVE_ACCOUNT,
@@ -50,17 +57,24 @@ export enum POPUPS_NUMBER {
   MANAGE_PROFILE
 }
 
-const AccountButton = ({ icon, onClick, text }: AccountButtonType) => {
+const AccountButton = ({ collapse = false, icon, onClick, text }: AccountButtonType) => {
   const theme = useTheme();
+
+  const collapsedStyle = collapse
+    ? {
+      '&:first-child': { '> span': { m: 0 }, m: '0px', minWidth: '48px' },
+      '> span': { m: 0 }
+    }
+    : {};
 
   return (
     <Button
       endIcon={icon}
       onClick={onClick}
-      sx={{ '&:hover': { bgcolor: 'divider' }, color: theme.palette.secondary.light, fontSize: '16px', fontWeight: 400, height: '53px', textTransform: 'none', width: 'fit-content' }}
+      sx={{ ...collapsedStyle, '&:hover': { bgcolor: 'divider' }, color: theme.palette.secondary.light, fontSize: '16px', fontWeight: 400, height: '53px', minWidth: '48px', textTransform: 'none', width: 'fit-content' }}
       variant='text'
     >
-      {text}
+      {collapse ? '' : text}
     </Button>
   );
 };
@@ -79,7 +93,8 @@ const AccountTotal = ({ hideNumbers, totalBalance }: { hideNumbers: boolean | un
           hideNumbers || hideNumbers === undefined
             ? <Box component='img' src={(theme.palette.mode === 'dark' ? stars6White : stars6Black) as string} sx={{ height: '36px', width: '154px' }} />
             : <FormatPrice
-              fontSize='32px'
+              commify
+              fontSize='24px'
               fontWeight={700}
               num={totalBalance}
               skeletonHeight={28}
@@ -92,14 +107,44 @@ const AccountTotal = ({ hideNumbers, totalBalance }: { hideNumbers: boolean | un
 };
 
 function AccountInformationForHome ({ accountAssets, address, hideNumbers, isChild, selectedAsset, setSelectedAsset }: AddressDetailsProps): React.ReactElement {
+  const nftManager = useMemo(() => new NftManager(), []);
+
   const { t } = useTranslation();
   const theme = useTheme();
   const pricesInCurrencies = usePrices();
   const currency = useCurrency();
   const account = useAccount(address);
-  const onAction = useContext(ActionContext);
 
   const [displayPopup, setDisplayPopup] = useState<number>();
+  const [myNfts, setNfts] = useState<ItemInformation[] | null | undefined>();
+
+  useEffect(() => {
+    if (!address) {
+      return;
+    }
+
+    // Handle updates after initialization
+    const handleNftUpdate = (updatedAddress: string, updatedNfts: ItemInformation[]) => {
+      if (updatedAddress === address) {
+        setNfts(updatedNfts);
+      }
+    };
+
+    // Waits for initialization
+    nftManager.waitForInitialization()
+      .then(() => {
+        setNfts(nftManager.get(address));
+      })
+      .catch(console.error);
+
+    // subscribe to the possible nft items for the account
+    nftManager.subscribe(handleNftUpdate);
+
+    // Cleanup
+    return () => {
+      nftManager.unsubscribe(handleNftUpdate);
+    };
+  }, [address, nftManager]);
 
   const calculatePrice = useCallback((amount: BN, decimal: number, price: number) => parseFloat(amountToHuman(amount, decimal)) * price, []);
 
@@ -133,10 +178,6 @@ function AccountInformationForHome ({ accountAssets, address, hideNumbers, isChi
     }).catch(console.error);
   }, [address, setSelectedAsset]);
 
-  const openSettings = useCallback((): void => {
-    address && onAction();
-  }, [onAction, address]);
-
   const goToDetails = useCallback((): void => {
     address && openOrFocusTab(`/accountfs/${address}/${selectedAsset?.assetId || '0'}`, true);
   }, [address, selectedAsset?.assetId]);
@@ -161,32 +202,42 @@ function AccountInformationForHome ({ accountAssets, address, hideNumbers, isChi
         <Grid container item justifyContent='flex-end' minHeight='50px'>
           <Divider sx={{ bgcolor: 'divider', height: '1px', mr: '5px', my: '15px', width: '100%' }} />
           <Grid container item xs>
-            {(assetsToShow === undefined || (assetsToShow && assetsToShow?.length > 0)) &&
-              <AOC
-                accountAssets={assetsToShow}
+            <Grid container item xs>
+              {(assetsToShow === undefined || (assetsToShow && assetsToShow?.length > 0)) &&
+                <AOC
+                  accountAssets={assetsToShow}
+                  address={address}
+                  hideNumbers={hideNumbers}
+                  mode='Home'
+                  onclick={onAssetBoxClicked}
+                  selectedAsset={selectedAsset}
+                />
+              }
+            </Grid>
+            <Grid container item width='fit-content'>
+              <NftGrouped
+                accountNft={myNfts}
                 address={address}
-                hideNumbers={hideNumbers}
-                mode='Home'
-                onclick={onAssetBoxClicked}
-                selectedAsset={selectedAsset}
               />
-            }
+            </Grid>
           </Grid>
           <Grid alignItems='center' container item width='fit-content'>
-            <Divider orientation='vertical' sx={{ bgcolor: 'divider', height: '34px', ml: 0, mr: '10px', my: 'auto', width: '1px' }} />
+            <Divider orientation='vertical' sx={{ bgcolor: 'divider', height: '34px', ml: 0, mr: '10px', mx: myNfts ? '5px' : undefined, my: 'auto', width: '1px' }} />
             <FullScreenAccountMenu
               address={address}
               baseButton={
                 <AccountButton
+                  collapse={!!myNfts}
                   icon={<MoreVertIcon style={{ color: theme.palette.secondary.light, fontSize: '32px' }} />}
-                  onClick={openSettings}
+                  onClick={noop}
                   text={t('Settings')}
                 />
               }
               setDisplayPopup={setDisplayPopup}
             />
-            <Divider orientation='vertical' sx={{ bgcolor: 'divider', height: '34px', ml: '5px', mr: '15px', my: 'auto', width: '1px' }} />
+            <Divider orientation='vertical' sx={{ bgcolor: 'divider', height: '34px', ml: '5px', mr: myNfts ? '5px' : '15px', my: 'auto', width: '1px' }} />
             <AccountButton
+              collapse={!!myNfts}
               icon={<ArrowForwardIosIcon style={{ color: theme.palette.secondary.light, fontSize: '28px' }} />}
               onClick={goToDetails}
               text={t('Details')}