+++ packages/extension-polkagate/src/hooks/useValidators.ts
@@ -3,7 +3,7 @@
 
 import type { Option } from '@polkadot/types';
 import type { AccountId } from '@polkadot/types/interfaces';
-//@ts-ignore
+// @ts-ignore
 import type { PalletStakingValidatorPrefs } from '@polkadot/types/lookup';
 import type { AnyJson } from '@polkadot/types/types';
 import type { BN } from '@polkadot/util';
@@ -105,7 +105,7 @@ export default function useValidators (address: string | undefined, validators?:
     }
 
     const getValidatorsPaged = async (eraIndex: number) => {
-      if (!api) {
+      if (!api || !currentEraIndex) {
         return; // never happens since we check api before, but to suppress linting
       }
 
@@ -152,11 +152,14 @@ export default function useValidators (address: string | undefined, validators?:
       const current: ValidatorInfo[] = [];
       const waiting: ValidatorInfo[] = [];
 
-      Object.keys(validatorPrefs).forEach((v) => {
-        Object.keys(currentEraValidatorsOverview).includes(v)
-          ? current.push(
+      for (const v of Object.keys(validatorPrefs)) {
+        if (Object.keys(currentEraValidatorsOverview).includes(v)) {
+          // const apy = await getValidatorApy(api, v, currentEraValidatorsOverview[v].total, validatorPrefs[v].commission, currentEraIndex);
+
+          current.push(
             {
               accountId: v as unknown as AccountId,
+              // apy,
               exposure: {
                 ...currentEraValidatorsOverview[v],
                 others: currentNominators[v]
@@ -164,8 +167,9 @@ export default function useValidators (address: string | undefined, validators?:
               stashId: v as unknown as AccountId,
               validatorPrefs: validatorPrefs[v]
             } as unknown as ValidatorInfo // types need to be revised!
-          )
-          : waiting.push(
+          );
+        } else {
+          waiting.push(
             {
               accountId: v as unknown as AccountId,
               exposure: {
@@ -177,7 +181,9 @@ export default function useValidators (address: string | undefined, validators?:
               validatorPrefs: validatorPrefs[v]
             } as unknown as ValidatorInfo
           );
-      });
+        }
+      }
+
       const inf = {
         current,
         eraIndex,