+++ packages/extension-polkagate/src/themes/light.ts
@@ -14,7 +14,7 @@ declare module '@mui/material/styles' {
     label: Palette['primary'];
     approval: Palette['primary'];
     support: Palette['primary'];
-    backgroundFL?: Palette['primary'];
+    backgroundFL: TypeText;
   }
   interface PaletteOptions {
     approval?: PaletteOptions['primary'];
@@ -41,7 +41,7 @@ export const lightTheme: ThemeOptions = {
     backgroundFL: { primary: '#DFDFDF', secondary: '#F1F1F1' },
     text: { primary: '#63364D', secondary: '#FFFFFF', disabled: '#747474' },
     action: { disabled: '#fff', disabledBackground: '#989898', focus: '#BA82A5' },
-    success: { main: '#1F7720' },
+    success: { main: '#1F7720', light: '#46890C' },
     warning: { main: '#FF002B' },
     divider: 'rgba(0, 0, 0, 0.1)'
   },